+++ src/lib/core/app.ts
@@ -2,9 +2,8 @@ import * as http from 'http'
 import Debug from 'debug'
 import { BlobTree } from '../storage/BlobTree'
 import { parseHttpRequest, WacLdpTask } from '../api/http/HttpParser'
-import { sendHttpResponse, WacLdpResponse, ErrorResult } from '../api/http/HttpResponder'
+import { sendHttpResponse, WacLdpResponse } from '../api/http/HttpResponder'
 import { executeTask } from './executeTask'
-import { makeResourceData, bufferToStream } from '../rdf/ResourceDataUtils'
 
 const debug = Debug('app')
 
@@ -30,20 +29,7 @@ export function makeHandler (storage: BlobTree, aud: string, skipWac: boolean) {
   return handle
 }
 
-export function makeRootAcl (owner: string): ReadableStream {
-  const obj = makeResourceData('text/turtle', [
-    `@prefix acl: <http://www.w3.org/ns/auth/acl#>.`,
-    `<#owner>`,
-    `  a acl:Authorization;`,
-    `  acl:agent <${owner}>;`,
-    `  acl:accessTo </>;`,
-    `  acl:default </>;`,
-    `  acl:mode`,
-    `    acl:Read, acl:Write, acl:Control.`
-  ].join('\n'))
-  const buffer = Buffer.from(JSON.stringify(obj))
-  return bufferToStream(buffer)
-}
+export { setRootAcl } from '../auth/setRootAcl'
 export { checkAccess, AccessCheckTask } from './checkAccess'
 export { determineWebId } from '../auth/determineWebId'
 export { BlobTree, Path } from '../storage/BlobTree'
